Rename GossipMessageDuplicateHandler -> GossipMessageDropHandler
